requirements.txt
aiocron==1.8
astropy==5.3.1
asyncio==3.4.3
beanie==1.21.0
black==23.7.0
bunnet==1.1.0
cachetools==5.3.1
dataclasses-json==0.5.14
discord-py-interactions==5.9.2
discord.py==2.3.2
httpx==0.24.1
jurigged==0.5.5
marshmallow-dataclass==8.5.14
python-blizzardapi==1.3.0
python-dotenv==1.0.0
typeguard==3.0.2
types-cachetools==5.3.0.6
wheel==0.41.1